BMS/requirement.txt
Django==2.1.7
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
uWSGI==2.0.18
